+++ crates/resolver/src/lookup.rs
@@ -35,26 +35,39 @@ use name_server_pool::{ConnectionHandle, ConnectionProvider, NameServerPool, Sta
 /// For IP resolution see LookupIp, as it has more features for A and AAAA lookups.
 #[derive(Clone, Debug, Eq, PartialEq)]
 pub struct Lookup {
+    query: Query,
     rdatas: Arc<Vec<RData>>,
     valid_until: Instant,
 }
 
 impl Lookup {
+    /// Return new instance with given rdata and the maximum TTL.
+    pub fn from_rdata(query: Query, data: RData) -> Self {
+        Self::new_with_max_ttl(query, Arc::new(vec![data]))
+    }
+
     /// Return new instance with given rdatas and the maximum TTL.
-    pub fn new_with_max_ttl(rdatas: Arc<Vec<RData>>) -> Self {
-        let valid_until = Instant::now() + Duration::from_secs(MAX_TTL as u64);
+    pub fn new_with_max_ttl(query: Query, rdatas: Arc<Vec<RData>>) -> Self {
+        let valid_until = Instant::now() + Duration::from_secs(u64::from(MAX_TTL));
         Lookup {
+            query,
             rdatas,
             valid_until,
         }
     }
 
     /// Return a new instance with the given rdatas and deadline.
-    pub fn new_with_deadline(rdatas: Arc<Vec<RData>>, valid_until: Instant) -> Self {
+    pub fn new_with_deadline(query: Query, rdatas: Arc<Vec<RData>>, valid_until: Instant) -> Self {
         Lookup {
+            query,
             rdatas,
             valid_until,
         }
+    }
+
+    /// Returns a reference to the Query that was used to produce this result.
+    pub fn query(&self) -> &Query {
+        &self.query
     }
 
     /// Returns a borrowed iterator of the returned IPs
@@ -75,6 +88,11 @@ impl Lookup {
         self.rdatas.len()
     }
 
+    #[cfg(test)]
+    pub fn rdatas(&self) -> &[RData] {
+        self.rdatas.as_ref()
+    }
+
     /// Clones the inner vec, appends the other vec
     pub(crate) fn append(&self, other: Lookup) -> Self {
         let mut rdatas = Vec::with_capacity(self.len() + other.len());
@@ -83,13 +101,7 @@ impl Lookup {
 
         // Choose the sooner deadline of the two lookups.
         let valid_until = min(self.valid_until(), other.valid_until());
-        Self::new_with_deadline(Arc::new(rdatas), valid_until)
-    }
-}
-
-impl From<RData> for Lookup {
-    fn from(data: RData) -> Self {
-        Lookup::new_with_max_ttl(Arc::new(vec![data]))
+        Self::new_with_deadline(self.query.clone(), Arc::new(rdatas), valid_until)
     }
 }
 
@@ -173,7 +185,7 @@ impl<C: DnsHandle + 'static> LookupFuture<C> {
         };
 
         LookupFuture {
-            client_cache: client_cache,
+            client_cache,
             names,
             record_type,
             options,
@@ -237,6 +249,11 @@ impl SrvLookup {
         SrvLookupIter(self.0.iter())
     }
 
+    /// Returns a reference to the Query that was used to produce this result.
+    pub fn query(&self) -> &Query {
+        self.0.query()
+    }
+
     /// Returns the list of IPs associated with the SRV record.
     ///
     /// *Note*: the lack of any IPs does not necessarily meant that there are no IPs available for the service, only that they were not included in the original request. A subsequent query for the IPs via the `srv.target()` should resolve to the IPs.
@@ -262,8 +279,7 @@ impl<'i> Iterator for SrvLookupIter<'i> {
         iter.filter_map(|rdata| match *rdata {
             RData::SRV(ref data) => Some(data),
             _ => None,
-        })
-        .next()
+        }).next()
     }
 }
 
@@ -301,6 +317,11 @@ macro_rules! lookup_type {
             pub fn iter(&self) -> $i {
                 $i(self.0.iter())
             }
+
+            /// Returns a reference to the Query that was used to produce this result.
+            pub fn query(&self) -> &Query {
+                self.0.query()
+            }
         }
 
         impl From<Lookup> for $l {
@@ -320,8 +341,7 @@ macro_rules! lookup_type {
                 iter.filter_map(|rdata| match *rdata {
                     $r(ref data) => Some(data),
                     _ => None,
-                })
-                .next()
+                }).next()
             }
         }
 
@@ -404,7 +424,7 @@ pub mod tests {
 
         fn send<R: Into<DnsRequest>>(&mut self, _: R) -> Self::Response {
             Box::new(future::result(
-                self.messages.lock().unwrap().pop().unwrap_or(empty()),
+                self.messages.lock().unwrap().pop().unwrap_or_else(empty),
             ))
         }
     }
@@ -442,8 +462,7 @@ pub mod tests {
                 RecordType::A,
                 DnsRequestOptions::default(),
                 CachingClient::new(0, mock(vec![v4_message()])),
-            )
-            .wait()
+            ).wait()
             .unwrap()
             .iter()
             .map(|r| r.to_ip_addr().unwrap())
@@ -454,14 +473,15 @@ pub mod tests {
 
     #[test]
     fn test_error() {
-        assert!(LookupFuture::lookup(
-            vec![Name::root()],
-            RecordType::A,
-            DnsRequestOptions::default(),
-            CachingClient::new(0, mock(vec![error()])),
-        )
-        .wait()
-        .is_err());
+        assert!(
+            LookupFuture::lookup(
+                vec![Name::root()],
+                RecordType::A,
+                DnsRequestOptions::default(),
+                CachingClient::new(0, mock(vec![error()])),
+            ).wait()
+            .is_err()
+        );
     }
 
     #[test]
@@ -472,8 +492,7 @@ pub mod tests {
                 RecordType::A,
                 DnsRequestOptions::default(),
                 CachingClient::new(0, mock(vec![empty()])),
-            )
-            .wait()
+            ).wait()
             .unwrap_err()
             .kind(),
             ResolveErrorKind::NoRecordsFound {