+++ crates/resolver/CHANGELOG.md
@@ -3,6 +3,16 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+## 0.11
+
+### Added
+
+- New option to execute queries concurrently, default is 2 #615
+
+### Changed
+
+- Added option to distrust Nameservers on SERVFAIL responses, continue resolution #613
+
 ## 0.10
 
 ### Fixed